+++ src/config.js
@@ -27,47 +27,47 @@ module.exports = {
   // ---npm packages
   npmPkgs: {
     eslint: {
-      eslint: '8.10.0',
-      'eslint-plugin-html': '6.2.0',
+      eslint: '8.21.0',
+      'eslint-plugin-html': '7.1.0',
     },
     prettier: {
-      prettier: '2.5.1',
+      prettier: '2.7.1',
       'eslint-config-prettier': '8.5.0',
-      'eslint-plugin-prettier': '4.0.0',
+      'eslint-plugin-prettier': '4.2.1',
     },
     js: {},
     jsx: {
-      '@babel/core': '7.18.2',
-      '@babel/eslint-parser': '7.18.2',
+      '@babel/core': '7.18.10',
+      '@babel/eslint-parser': '7.18.9',
     },
     ts: {
-      '@typescript-eslint/eslint-plugin': '5.14.0',
-      '@typescript-eslint/parser': '5.14.0',
+      '@typescript-eslint/eslint-plugin': '5.32.0',
+      '@typescript-eslint/parser': '5.32.0',
     },
     vue2: {
       'vue-eslint-parser': '9.0.3',
-      'eslint-plugin-vue': '9.1.1',
+      'eslint-plugin-vue': '9.3.0',
     },
     vue3: {
       'vue-eslint-parser': '9.0.3',
-      'eslint-plugin-vue': '9.1.1',
+      'eslint-plugin-vue': '9.3.0',
     },
     'vue2-ts': {
       'vue-eslint-parser': '9.0.3',
-      'eslint-plugin-vue': '9.1.1',
-      '@typescript-eslint/eslint-plugin': '5.14.0',
-      '@typescript-eslint/parser': '5.14.0',
+      'eslint-plugin-vue': '9.3.0',
+      '@typescript-eslint/eslint-plugin': '5.32.0',
+      '@typescript-eslint/parser': '5.32.0',
       '@vue/eslint-config-typescript': '8.0.0',
     },
     'vue3-ts': {
       'vue-eslint-parser': '9.0.3',
-      'eslint-plugin-vue': '9.1.1',
-      '@typescript-eslint/eslint-plugin': '5.14.0',
-      '@typescript-eslint/parser': '5.14.0',
+      'eslint-plugin-vue': '9.3.0',
+      '@typescript-eslint/eslint-plugin': '5.32.0',
+      '@typescript-eslint/parser': '5.32.0',
       '@vue/eslint-config-typescript': '8.0.0',
     },
     svelte3: {
-      'eslint-plugin-svelte3': '3.4.1',
+      'eslint-plugin-svelte3': '4.0.0',
     },
   },
   // platform - settings.json path