+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * add: version log
 * add: `chalk` highlight log
 * add: use `vue-eslint-parser` parser in (`vue2-ts` / `vue3-ts`) eslint config file
+* update: npm dependencies version
 ## 1.4.4
 * support vue2.7 setup macros `defineProps` `defineEmits` `defineExpose` `withDefaults`
 * upgrade dependencies `eslint-plugin-vue` `vue-eslint-parser`